requirements.txt
google-generativeai==0.3.2
cryptography==41.0.7
psutil==5.9.6
requests==2.31.0
numpy==1.24.3
pycuda
opencl
pyopencl
hashlib-compat
pycryptodome==3.19.0